⚡:: (#49) 불필요하게 파일 늘리지 않고 common에 모으기
