+++ src/components/common/write/style.ts
